+++ code/src/ui/interaction/form/section/form-section.tsx
@@ -21,12 +21,14 @@ interface FormSectionProps {
  */
 export default function FormSection(props: Readonly<FormSectionProps>) {
   return (
-    <div className="p-1 md:p-6 flex flex-col justify-center mx-auto border-1 border-border bg-background  rounded-lg m-4  ">
-      <h2 className=" text-lg md:text-2xl  font-bold">
+    <div className="p-2 md:p-6 flex flex-col justify-center mx-auto border-1 border-border bg-background rounded-lg m-4  ">
+      <h2 className=" text-xl md:text-2xl  font-bold">
         {parseWordsForLabels(props.group.label[VALUE_KEY])}
       </h2>
-      <div className="p-2 space-y-2">     
-        {!props.group.maxCount || (props.group.maxCount && parseInt(props.group.maxCount?.[VALUE_KEY]) > 1) ?
+      <div className="p-2 space-y-2">
+        {!props.group.maxCount ||
+        (props.group.maxCount &&
+          parseInt(props.group.maxCount?.[VALUE_KEY]) > 1) ? (
           <FormArray
             fieldId={props.group.label[VALUE_KEY]}
             minSize={parseInt(props.group.minCount?.[VALUE_KEY])}
@@ -34,9 +36,12 @@ export default function FormSection(props: Readonly<FormSectionProps>) {
             fieldConfigs={props.group.property}
             form={props.form}
             options={props.options}
-          /> : props.group.property.map((field, index) =>
-            renderFormField(props.entityType, field, props.form, index))
-        }
+          />
+        ) : (
+          props.group.property.map((field, index) =>
+            renderFormField(props.entityType, field, props.form, index)
+          )
+        )}
       </div>
     </div>
   );